agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psutil==5.9.8
pyyaml==6.0.1
cryptography==42.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
websockets==12.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
